feat: further refactor, add toast on success to not refresh page
