+++ src/lib/graphql/flatfile/client.ts
@@ -1,6 +1,6 @@
 import { GraphQLClient } from "graphql-request";
 
-import { FLATFILE_GRAPHQL_URL } from "src/constants/flatfile";
+import { FLATFILE_GRAPHQL_URL } from "src/constants/csv";
 
 export const createFlatfileClient = (token: string) =>
   new GraphQLClient(FLATFILE_GRAPHQL_URL, {
